project/requirements.txt
# ── Backend ──────────────────────────
fastapi>=0.110.1
uvicorn>=0.29.0
sqlalchemy>=2.0.29
psycopg2-binary>=2.9.9
pydantic>=2.6.4
python-dotenv>=1.0.1
python-multipart>=0.0.9

# ── OCR ──────────────────────────────
easyocr>=1.7.1
opencv-python>=4.9.0.80
pillow>=10.3.0
numpy>=1.26.0

# ── Frontend ─────────────────────────
streamlit>=1.32.2
requests>=2.31.0
pandas>=2.2.1
plotly>=5.20.0
